backend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-dotenv
httpx
pytz
requests
MetaTrader5
